NP-14149 Split into multiple files
