+++ src/pages/public_registration/PublicRelatedResourcesContent.tsx
@@ -1,5 +1,5 @@
 import { CircularProgress, Link, Typography } from '@mui/material';
-import { useCallback, useEffect, useState } from 'react';
+import { useCallback, useEffect, useMemo, useState } from 'react';
 import { apiRequest } from '../../api/apiRequest';
 import { RegistrationList } from '../../components/RegistrationList';
 import { Registration } from '../../types/registration.types';
@@ -10,7 +10,7 @@ interface PublicRelatedResourcesContentProps {
 }
 
 export const PublicRelatedResourcesContent = ({ related = [] }: PublicRelatedResourcesContentProps) => {
-  const internalResources = related.filter((uri) => uri.includes(API_URL));
+  const internalResources = useMemo(() => related.filter((uri) => uri.includes(API_URL)), [related]);
   const externalResources = related.filter((uri) => !uri.includes(API_URL));
 
   const [isLoadingRegistrations, setIsLoadingRegistrations] = useState(true);
@@ -29,7 +29,7 @@ export const PublicRelatedResourcesContent = ({ related = [] }: PublicRelatedRes
     ) as Registration[];
     setRelatedRegistrations(registrations);
     setIsLoadingRegistrations(false);
-  }, []); // todo: infinite loop when using array as dependancy?
+  }, [internalResources]);
 
   useEffect(() => {
     getRelatedRegistrations();
@@ -37,20 +37,16 @@ export const PublicRelatedResourcesContent = ({ related = [] }: PublicRelatedRes
 
   return (
     <>
-      {internalResources.length > 0 && (
-        <>
-          <Typography variant="overline">Interne</Typography>
-          {isLoadingRegistrations ? (
-            <CircularProgress aria-labelledby="TODO" sx={{ display: 'block' }} />
-          ) : (
-            <RegistrationList registrations={relatedRegistrations} />
-          )}
-        </>
-      )}
+      {internalResources.length > 0 &&
+        (isLoadingRegistrations ? (
+          <CircularProgress aria-labelledby="TODO" sx={{ display: 'block' }} />
+        ) : (
+          <RegistrationList registrations={relatedRegistrations} />
+        ))}
 
       {externalResources.length > 0 && (
         <>
-          <Typography variant="overline">Eksterne</Typography>
+          <Typography variant="overline">Eksterne lenker</Typography>
           <ul>
             {externalResources.map((externalResource) => (
               <li>
